// src/components/verifi-vote/types.ts
export type NetworkID =
  | 's'
  | 's-tn'
  | 'eth'
  | 'gor'
  | 'sep'
  | 'linea-testnet'
  | 'sn'
  | 'sn-tn'
  | 'matic'
  | 'arb1'
  | 'magma-tn';
export type Choice = 1 | 2 | 3;


type ChainAsNetworkID = {
  [chainId: number]: NetworkID;
};

export const chainAsNetworkID: ChainAsNetworkID = {
  // Assign chain IDs to their corresponding NetworkID
  1: 'eth',
  3: 'gor',
  137: 'matic',
  42161: 'arb1',
  // ... add more mappings as needed
};

export type SelectedStrategy = {
  address: string;
  type: string;
};

export type SpaceMetadataDelegation = {
  name: string | null;
  apiType: string | null;
  apiUrl: string | null;
  contractNetwork: NetworkID | null;
  contractAddress: string | null;
};
// ...
export type StrategyParsedMetadata = {
  name: string;
  description: string;
  decimals: number;
  symbol: string;
  token: string | null;
  payload: string | null;
};

export type Space = {
  id: string;
  network: NetworkID;
  name: string;
  avatar: string;
  cover: string;
  about?: string;
  external_url: string;
  delegations: SpaceMetadataDelegation[];
  twitter: string;
  github: string;
  discord: string;
  voting_power_symbol: string;
  wallet: string;
  controller: string;
  voting_delay: number;
  min_voting_period: number;
  max_voting_period: number;
  proposal_threshold: string;
  validation_strategy: string;
  validation_strategy_params: string;
  voting_power_validation_strategy_strategies: string[];
  voting_power_validation_strategy_strategies_params: string[];
  voting_power_validation_strategies_parsed_metadata: StrategyParsedMetadata[];
  strategies_indicies: number[];
  strategies: string[];
  strategies_params: any[];
  strategies_parsed_metadata: StrategyParsedMetadata[];
  authenticators: string[];
  executors: string[];
  executors_types: string[];
  proposal_count: number;
  vote_count: number;
  created: number;
};